+++ app/src/main/java/com/tinkoff/edu/app/LoanRequest.java
@@ -1,5 +1,7 @@
 package com.tinkoff.edu.app;
 
+import com.tinkoff.edu.app.dictionary.LoanType;
+
 /**
  * Class, Type -> objects, instances
  */
