+++ app/src/main/java/com/tinkoff/edu/app/dictionary/LoanDecision.java
@@ -1,4 +1,4 @@
-package com.tinkoff.edu.app;
+package com.tinkoff.edu.app.dictionary;
 
 /**
  * Type of bank answer
