+++ app/src/main/java/com/tinkoff/edu/app/dictionary/LoanType.java
@@ -1,4 +1,4 @@
-package com.tinkoff.edu.app;
+package com.tinkoff.edu.app.dictionary;
 
 public enum LoanType {
     PERSON,
