+++ app/src/main/java/com/tinkoff/edu/app/repository/LoanCalcRepository.java
@@ -1,0 +1,8 @@
+package com.tinkoff.edu.app.repository;
+
+import com.tinkoff.edu.app.LoanRequest;
+import com.tinkoff.edu.app.LoanResponse;
+
+public interface LoanCalcRepository {
+    LoanResponse save(LoanRequest request);
+}
