+++ app/src/main/java/com/tinkoff/edu/app/service/LoanCalcService.java
@@ -1,0 +1,8 @@
+package com.tinkoff.edu.app.service;
+
+import com.tinkoff.edu.app.LoanRequest;
+import com.tinkoff.edu.app.LoanResponse;
+
+public interface LoanCalcService {
+    LoanResponse createRequest(LoanRequest request);
+}
