+++ app/src/main/java/com/tinkoff/edu/test/LoanCalcTest.java
@@ -1,9 +1,11 @@
 package com.tinkoff.edu.test;
 
-import com.tinkoff.edu.app.LoanCalcController;
 import com.tinkoff.edu.app.LoanRequest;
-import com.tinkoff.edu.app.LoanType;
-import com.tinkoff.edu.app.StaticVariableLoanCalcRepository;
+import com.tinkoff.edu.app.controller.LoanCalcController;
+import com.tinkoff.edu.app.dictionary.LoanType;
+import com.tinkoff.edu.app.repository.StaticVariableLoanCalcRepository;
+import com.tinkoff.edu.app.service.IpNotFriendlyService;
+import com.tinkoff.edu.app.service.LoanCalcService;
 
 
 /**
@@ -12,7 +14,8 @@ import com.tinkoff.edu.app.StaticVariableLoanCalcRepository;
 public class LoanCalcTest {
     public static void main(String... args) {
         LoanRequest request = new LoanRequest(10, 1000, LoanType.IP);
-        LoanCalcController loanCalcController = new LoanCalcController(new StaticVariableLoanCalcRepository());
+        LoanCalcService loanCalcService = new IpNotFriendlyService(new StaticVariableLoanCalcRepository());
+        LoanCalcController loanCalcController = new LoanCalcController(loanCalcService);
 
         int requestId = loanCalcController.createRequest(request).getRequestId();
